added stepper and switched to constant enums
