+++ packages/web-app/src/utils/constants.ts
@@ -5,7 +5,7 @@ export const DEFAULT_CURRENCY = 'usd';
 export const INFURA_PROJECT_ID = '7a03fcb37be7479da06f92c5117afd47';
 
 /** Time period options for token price change */
-export enum TimeFilter {
+export const enum TimeFilter {
   day = 'day',
   week = 'week',
   month = 'month',
@@ -20,7 +20,7 @@ export const NAV_LINKS = [
   {label: 'Community', path: Community},
 ];
 
-export enum TransferTypes {
+export const enum TransferTypes {
   Deposit = 'Deposit',
   Withdraw = 'Withdraw',
 }
